+++ nas-20170626/alibabacloud_nas20170626/models/_describe_file_systems_request.py
@@ -21,35 +21,53 @@ class DescribeFileSystemsRequest(DaraModel):
     ):
         # The file system ID.
         # - General-purpose NAS: 31a8e4****.
-        # - Extreme NAS: must start with extreme-, such as extreme-0015****.
-        # - CPFS (locally redundant): must start with cpfs-, such as cpfs-125487****.
-        # - CPFS SE (zone-redundant): must start with cpfsse-, such as cpfsse-022c71b134****.
+        # - Extreme NAS: Must start with extreme-, such as extreme-0015****.
+        # - Cloud Parallel File Storage (CPFS) (locally redundant version): Must start with cpfs-, such as cpfs-125487****.
+        # - CPFS SE (zone-redundant version): Must start with cpfsse-, such as cpfsse-022c71b134****.
         self.file_system_id = file_system_id
-        # The type of the file system.
+        # The file system type.
         # 
         # Valid values:
         # 
         # - all (default): queries all types.
         # - standard: General-purpose NAS.
         # - extreme: Extreme NAS.
-        # - cpfs: Cloud Parallel File Storage (CPFS) with locally redundant storage.
-        # - cpfsse: CPFS SE with zone-redundant storage.
+        # - cpfs: Cloud Parallel File Storage (CPFS) (locally redundant version).
+        # - cpfsse: CPFS SE (zone-redundant version).
         # 
         # > To query multiple types, separate them with commas (,).
         self.file_system_type = file_system_type
         # The page number of the file system list.
+        # 
+        # Default value: 1.
         self.page_number = page_number
-        # The number of file systems on each page in a paging query.
+        # The number of file systems on each page during a paged query.
+        # 
+        # Valid values: 1 to 100.
+        # 
+        # Default value: 10.
         self.page_size = page_size
         # The resource group ID.
         # 
         # You can view the resource group ID in the [Resource Management console](https://resourcemanager.console.aliyun.com/resource-groups?).
         self.resource_group_id = resource_group_id
-        # The storage type. Currently, only CPFS for Lingjun specifications are supported for a filtered query. Other FileSystemType values are not supported. The following specifications are supported:
+        # The storage type. Currently, only CPFS for Lingjun specifications are supported for filtered query. Other file system types (FileSystemType) are not supported. The following specifications are supported:
+        # - bm_advance_400
+        # - bm_advance_400_large
+        # - bm_advance_new
+        # - bm_advance_100
+        # - cpfslight_100
+        # - bm_p100_stepsize_300
+        # - bm_p400_stepsize_1
+        # - bm_p400_stepsize_300
+        # 
+        # You can specify multiple specifications separated by commas (,). For example: bm_advance_400,bm_advance_new.
         self.storage_type = storage_type
         # The tag information.
         self.tag = tag
-        # The VPC ID.
+        # The virtual private cloud (VPC) ID.
+        # 
+        # The VPC must be the same as the VPC of the Elastic Computing Service (ECS) server to which you want to mount the file system.
         self.vpc_id = vpc_id
 
     def validate(self):
@@ -141,7 +159,7 @@ class DescribeFileSystemsRequestTag(DaraModel):
         # 
         # Limits:
         # 
-        # - N can be an integer from 1 to 20.
+        # - Valid values of N: 1 to 20.
         # - The tag value can be up to 128 characters in length.
         # - The tag value cannot start with `aliyun` or `acs:`.
         # - The tag value cannot contain `http://` or `https://`.